move CLProgram to own file
